add deriver, updater libraries to process
